增加视频

+++ videos/views.py
@@ -42,29 +42,12 @@ def upload_video(request):
         label = request.POST.get('label')
         title = request.POST.get('title')
         description = request.POST.get('description')
-        max_id = Video.objects.all().aggregate(Max('id'))['id__max']
-        video_id = max_id + 1 if max_id else 1
-        
-        # # 将视频和封面文件上传到云服务器
-        # video_storage = SFTPStorage()
-        # cover_storage = SFTPStorage()
-        # video_storage.location = 'data/video_file/'
-        # cover_storage.location = 'data/video_cover/'
-        # print("video_storage : ",video_storage)
-        # print("cover_storage_storage : ",cover_storage)
-        # print("video_storage.location : ",video_storage.location)
-        # print("cover_storage.location : ",cover_storage.location)
-        # #save() 方法的参数是 Django 的 File 对象
-        # #upload() 方法的参数是本地文件路径
-        # video_path = video_storage.save(f'{video_id}_{title}.mp4', video_file)
-        # cover_path = cover_storage.save(f'{video_id}_{title}.png', cover_file)
-        # print("video_path : ",video_path)
-        # print("cover_path : ",cover_path)
-        # video_storage.close()
-        # cover_storage.close()
+        video = Video.objects.filter(title=title)
+        if video.exists():
+            return JsonResponse({'errno': 2003, 'msg': "视频名称重复！"})
 
-        video_path = os.path.join(settings.VIDRO_URL, f'{video_id}_{title}.mp4')
-        cover_path = os.path.join(settings.COVER_URL, f'{video_id}_{title}.png')
+        video_path = os.path.join(settings.VIDRO_URL, f'{title}.mp4')
+        cover_path = os.path.join(settings.COVER_URL, f'{title}.png')
 
         with open(video_path, 'wb+') as f:
             for chunk in video_file.chunks():
